// src/libs/models/player.ts
import { FocusNode } from "./models";
import { Car } from "./car";
import { Material } from "./material";

/**
 * Miata Rpg API
 * This is a simple API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Player { 
    name: string;
    description: string;
    avatar: string;
    color: string;
    car : Car;
    talents: Array<PlayerTalent>;
    materials: Array<Material>
}

export interface PlayerTalent { 
    name: string;
    effect: number;
    focus: FocusNode;
}