// src/libs/models/material.ts
/**
 * Miata Rpg API
 * This is a simple API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReferralObject } from './referralObject';

export interface Material { 
    title: string;
    description: string;
    referral: ReferralObject;
    cost: number;
    qty: number;
}